Remove notification_views_map from MessageCenterView
This patch removed the cache (notification_views_map) from MessageCenterView. Instead, it refers MessageListView directly.

As the investigation (see crbug.com/737858#3), the cache and the list view are sometimes inconsistent. To remove the inconsistency, this patch removes the cache. Then, the Issue 737858 should be fixed by this patch.

Bug: 737858
Test: manual

+++ ash/message_center/message_center_view.h
@@ -74,7 +74,7 @@ class ASH_EXPORT MessageCenterView
   void OnDidChangeFocus(views::View* before, views::View* now) override;
 
   static const SkColor kBackgroundColor;
-  static const size_t kMaxVisibleNotifications;
+  static const int kMaxVisibleNotifications;
 
  protected:
   // Potentially sets the reposition target, and then returns whether or not it
